Fix negative canvas_print return value by initializing total_write

+++ src/canvas.h
@@ -36,7 +36,7 @@ PIXEL *canvas_get_unck(struct canvas *self, size_t x, size_t y);
 
 PIXEL *canvas_get(struct canvas *self, size_t x, size_t y);
 
-long canvas_print(const struct canvas *self, FILE *f);
+int canvas_print(const struct canvas *self, FILE *f);
 
 int canvas_unprint(const struct canvas *self, FILE *f);
